# openapi.yaml
openapi: 3.1.0
info:
  title: E-commerce Polyglot API
  version: 1.0.0
  description: API para plataforma de comercio electrónico con persistencia poliglota
  contact:
    name: Equipo de Desarrollo
    url: https://github.com/yourusername/ING2_UADE_TPO
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderItem:
      type: object
      required:
        - product
        - quantity
        - price
      properties:
        product:
          type: string
          format: mongodb-objectid
        quantity:
          type: integer
          minimum: 1
        price:
          type: number
          minimum: 0
    ShippingAddress:
      type: object
      required:
        - street
        - city
        - state
        - zipCode
        - country
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
    PaymentInfo:
      type: object
      required:
        - method
        - transactionId
      properties:
        method:
          type: string
          enum: [credit_card, debit_card, transfer]
        transactionId:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
          default: pending
    Order:
      type: object
      required:
        - user
        - items
        - total
        - shippingAddress
        - paymentInfo
      properties:
        _id:
          type: string
          format: mongodb-objectid
        user:
          type: string
          format: mongodb-objectid
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        total:
          type: number
          minimum: 0
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
          default: pending
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        paymentInfo:
          $ref: '#/components/schemas/PaymentInfo'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    OrderAnalytics:
      type: object
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        averageOrderValue:
          type: number
        ordersByStatus:
          type: object
          additionalProperties:
            type: integer
        period:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
...
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: Listar pedidos
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, processing, shipped, delivered, cancelled]
      responses:
        200:
          description: Lista paginada de pedidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: integer
                  page:
                    type: integer
                  totalPages:
                    type: integer
                  hasMore:
                    type: boolean
        401:
          description: No autorizado

    post:
      tags: [Orders]
      summary: Crear un nuevo pedido
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - shippingAddress
                - paymentInfo
              properties:
                shippingAddress:
                  $ref: '#/components/schemas/ShippingAddress'
                paymentInfo:
                  $ref: '#/components/schemas/PaymentInfo'
      responses:
        201:
          description: Pedido creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Datos inválidos o carrito vacío
        401:
          description: No autorizado
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Obtener detalles de un pedido
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Detalles del pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        401:
          description: No autorizado
        403:
          description: No tiene permisos para ver este pedido
        404:
          description: Pedido no encontrado
  /orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Actualizar estado de un pedido
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [pending, processing, shipped, delivered, cancelled]
      responses:
        200:
          description: Estado actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Estado inválido
        401:
          description: No autorizado
        403:
          description: No tiene permisos de administrador
        404:
          description: Pedido no encontrado
  /orders/{id}/payment:
    patch:
      tags: [Orders]
      summary: Actualizar estado de pago de un pedido
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [pending, completed, failed]
      responses:
        200:
          description: Estado de pago actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Estado inválido
        401:
          description: No autorizado
        403:
          description: No tiene permisos de administrador
        404:
          description: Pedido no encontrado
  /orders/analytics:
    get:
      tags: [Orders]
      summary: Obtener análisis de pedidos
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
      responses:
        200:
          description: Análisis de pedidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderAnalytics'
        401:
          description: No autorizado
        403:
          description: No tiene permisos de administrador 